move monitoring_utils
